+++ src/main/java/at/hannibal2/skyhanni/data/model/ComposterUpgrade.kt
@@ -1,11 +1,11 @@
 package at.hannibal2.skyhanni.data.model
 
-enum class ComposterUpgrade(val displayName: String) {
-    COMPOSTER_SPEED("Composter Speed"),
-    ORGANIC_MATTER_CAP("Organic Matter Cap"),
-    MULTI_DROP("Multi Drop"),
-    COST_REDUCTION("Cost Reduction"),
-    FUEL_CAP("Fuel Cap"),
+enum class ComposterUpgrade(val displayName: String, val slotNumber: Int) {
+    COMPOSTER_SPEED("Composter Speed", 20),
+    MULTI_DROP("Multi Drop", 21),
+    FUEL_CAP("Fuel Cap", 22),
+    ORGANIC_MATTER_CAP("Organic Matter Cap", 23),
+    COST_REDUCTION("Cost Reduction", 24),
     ;
 
     companion object {