+++ src/main/java/at/hannibal2/skyhanni/features/bazaar/BazaarDataHolder.kt
@@ -17,7 +17,6 @@ class BazaarDataHolder {
     }
 
     private fun loadNpcPrices(): MutableMap<String, Double> {
-        println("loadNpcPrices")
         val list = mutableMapOf<String, Double>()
         try {
             val itemsData = APIUtil.getJSONResponse("https://api.hypixel.net/resources/skyblock/items")
@@ -55,7 +54,9 @@ class BazaarDataHolder {
         val buyPrice = NEUItems.getPrice(internalName, false)
         val npcPrice = npcPrices[internalName].let {
             if (it == null) {
-                LorenzUtils.debug("NPC price not found for item '$internalName'")
+                if (!ignoreNoNpcPrice(internalName)) {
+                    LorenzUtils.debug("NPC price not found for item '$internalName'")
+                }
                 0.0
             } else it
         }
@@ -63,5 +64,14 @@ class BazaarDataHolder {
         val data = BazaarData(internalName, displayName, sellPrice, buyPrice, npcPrice)
         bazaarData[internalName] = data
         return data
+    }
+
+    private fun ignoreNoNpcPrice(internalName: String): Boolean {
+        if (internalName.startsWith("TURBO_")) return true
+        if (internalName == "PURPLE_CANDY") return true
+        if (internalName == "JACOBS_TICKET") return true
+        if (internalName == "RAW_SOULFLOW") return true
+
+        return false
     }
 }