+++ CHANGELOG.md
@@ -67,6 +67,10 @@
 + Added command **/shcropspeedmeter** - Helps calculate the real farming fortune with the formula crops broken per block.
 + Added **Compost Low Notification** - Shows a notification as title when organic matter/fuel is low.
 + Added **Jacob's Contest Warning** - Show a warning shortly before a new jacob contest starts.
++ Added **Inventory Numbers** - Show the number of the teleport pads inside the 'Change Destination' inventory as stack size.
++ Added **Composter Overlay** - Show the cheapest items for organic matter and fuel, show profit per compost/hour/day and time per compost
++ Added **Composter Upgrades Overlay** - Show an overview of all composter stats, including time till organic matter and fuel is empty when fully filled and show a preview how these stats change when hovering over an upgrade
++ Hide crop money display, crop milestone display and garden visitor list while inside anita show, SkyMart or the composter inventory
 
 ### Features from other Mods
 > *The following features are only there because I want them when testing SkyHanni features without other mods present.*